fixup: Removed poorly implemented and unused constant index operator
We cannot use non-constant raw() to implement a constant index operator.
The compiler probably does not complain because that constant operator
is currently unused (and belongs to a templated class). We could add a
constant raw(), but doing that properly requires renaming both raw() to
start(), making them private and providing an indexOf() replacement for
the only external raw() caller use case. All that work is outside this
build fix scope.