Document MemObject::endOffset()
It is difficult to document the exact meaning of endOffset(), without
going into many details, because, when dealing with HTTP 206 (Partial
Content) responses, Squid sometimes stores their bodies using true
Range-specific offsets and sometimes treats them as if they were HTTP
200 (OK) responses without ranges.

For example, a one-part 206 response to a forwarded
Range:bytes=1000-1000 client request may be stored with the first body
byte offset of 1300, where 300 is the size of the corresponding
serialized HTTP 206 response header. However, a multipart response with
an identical first part is stored with the first body byte offset of 300
because Squid will not "do ranges" and just pretend it is a 200-like
(uncachable) response!