fixup: Replaced problematic netdb MessageSizeMax with out-of-scope TODO
The official code does not limit. We are dealing with a probably trusted
cache_peer so while it may not be acceptable to let it run us out of
memory, we do not have to add those protections in this branch.

Our initial attempt to limit was problematic: We should check the limit
_after_ we parsed data to avoid _requesting_ bytes that do not want to
parse (rather than rejecting bytes that we ourselves have asked for).

The correct limit itself would be difficult to establish here because
these netdb exchanges can be quite large in some environments. It may
need to be configurable.