Started converting URN code to use Store::ParsingBuffer
It looks like we can easily avoid zero-termination issues because the
official urnParseReply() already re-allocates its input buffer. We just
cannot use xstrndup() because that abomination of a wrapper requires
zero termination of its argument.

I hope that temporary hackBufferXXX hack will be easily removed when we
merge Store::ReadBuffer and Store::ParsingBuffer, which is currently the
plan.