fixup: Fix EOF detection via store_client::atEof() callers
While this change probably fixes buggy atEof() implementation that I
added in branch commit d5d8c40, I am still not fully satisfied with this
solution because it stores information that should be a part of
store_client's answer in the store_client object itself.

N.B. The primary code path (SendMoreData() and friends) does not detect
Store EOF based on the store_client response. Instead, after handling
HTTP headers, it passes Store body data buffer to Client Streams, and
Client Streams do not know about Store (and can use a simple
!StoreIOBuffer.length check).

We could refactor all secondary STCB callbacks to handle headers
separately, but that would either require splitting those callbacks into
header-only and body-only callbacks (like primary code does with
CacheHit and HandleIMSReply) or counting the number of Store answers.
The added store_client::atEof() effectively does the latter, reducing
code duplication.

We could add the flag as a third STCB parameter. However, this will
introduce a lot of noise into the diff and the parameter will be ignored
by two primary code path callbacks (CacheHit, HandleIMSReply) because
they only deal with HTTP headers. Adding an HttpReply pointer would be
arguably worse because it would add a _third_ response object for the
recipient to pick from (MemObject::baseReply() and
MemObject::freshestReply() being the other two).

We could also re-add the EOF flag to StoreIOBufffer, but that is the
worst approach out of all because a lot of Client Streams and other
StoreIOBuffer-using code does not need that flag and will not know what
to do with it. We already have similar problems with the error flag...