Handle unparseable WHOIS AS responses that fill the entire read buffer
My "unparsed bytes" part of the "stop suspicious accumulation of ..."
claim in earlier branch code was wrong: The stillReasonableOffset check
does not prevent the read buffer from becoming full, potentially causing
an infinite read loop. The claim would have been correct if we were to
grow the buffer space, but we do not do that (yet?).