fixup: Fix sendClientUpstreamResponse() path
    ERROR: Squid BUG: assurance failed: result.data == next()->readBuffer.data
    exception location: client_side_reply.cc(2046) noteStreamBufferredBytes

The upstream response is delivered via a special
clientReplyContext::tempbuf buffer rather than the usual
next()->readBuffer. Official code has a special `if` statement to copy
that special buffer to the usual one and a special pushStreamData()
parameter to supply the right/usual data pointer with the special
answer.

My new noteStreamBufferredBytes() code missed this special case. We
could add a special parameter like pushStreamData() had or limit the
special case to that `if` statement. The latter is a much cleaner/safer
solution IMO.