fixup: Fix branch handling of certain truncated responses
It turns out that some official code insists that a zero-length
StoreIOBuffer also has a nil data pointer:

    !receivedData.data &&
    !receivedData.length;

    receivedData.data == nullptr && receivedData.length == 0

We should relax those checks (and move them into a StoreIOBuffer bool
conversion operator), but they are located in seemingly unrelated code,
so that improvement is probably done best in a dedicated branch. For
now, we try to create the StoreIOBuffer objects the above code expects.

This also fixes Store::ParsingBuffer::appended() assertions that
incorrectly ~relied on memory() never being nil. The branch code did not
actually guarantee that readerSuppliedMemory_.data was never nil.