fixup: Do not assert at the end of the data on some range requests
    client_side_reply.cc:1736: "result.offset == next()->readBuffer.offset"

On empty storeClientCopy() result, legacy clientReplyContext calls
sendStreamError() instead of pushStreamData(). IMO, that is a bug -- the
end of stored body is not an error, but that legacy code "works",
probably because both methods end up calling clientStreamCallback() with
nearly identical parameters, resulting in nearly identical state. The
only material difference that I know about is that sendStreamError()
sets http->request->flags.streamError flag (which, again, feels wrong
when dealing with a benign error-free EOF case.

The branch code triggered the assertion quoted above because we did not
bother to set result.offset for empty storeClientCopy() results. Now we
do set it.

I also relaxed the assertion itself because it was too restrictive -- an
empty content range cannot have a meaningful offset. There are already
similar "relaxed" assertions in the official code, but there are also
some that look strict. I cannot easily tell whether the latter code is
correct; its preconditions/context is not clear to me.

When all code stops checking offsets of empty buffers (and stops
insisting on empty buffers having nil data pointers), we can remove the
hacks/workarounds/preconditions added or adjusted in this commit.