fixup: Do not downgrade log level for MemStore (parsing) errors
It is questionable that these (and similar cache_dir) errors should be
reported at level 1, as discussed in
store_client::parseHttpHeadersFromDisk(), but perhaps this branch is not
the best place to answer that question. Let's preserve the log level of
the official code (even though we are potentially expanding the number
of problems that might be logged now).