fixup: Reject bad storeClientCopy() calls at build time
Most storeClientCopy() calls inside clientReplyContext must invalidate
lastStreamBufferedBytes. The current code does this correctly, but new
code might forget (and backports may miss this dependency). Thus, it is
a good idea to prohibit such "accidental" storeClientCopy() calls.

A simpler approach would just duplicate the current ::storeClientCopy()
signature to add this static function member:

```C++
    // keep in sync with ::storeClientCopy() declaration!
    static void storeClientCopy(store_client *, StoreEntry *, StoreIOBuffer, STCB *, void *) = delete;
```

However, when ::storeClientCopy() signature changes, our protection will
silently stop working. Unfortunately, a static assert cannot check that
the types of global storeClientCopy() and our deleted storeClientCopy()
member are the same because decltype() are not allowed to use deleted
functions. We would have to rely on manual syncing if we delete this
function member explicitly.

Ideally, we want this:

```C++
    static decltype(::storeClientCopy) storeClientCopy = delete;
```

but that just does not compile:

    error: expected primary-expression before ‘;’ token