Add farproc_cast template
On Windows FARPROC is defined as 'long long int (*)()
however is it treated as a (void *)() function pointer
by the GetProcAddress() API for DLL symbol loading.

This non-void type for FARPROC triggers the GCC
-Wcast-function-type protection designed to prevent
casting function pointers with incompatible stack
layouts.

Provide a template that performs the correct function
pointer reinterpret_cast with that specific diagnostic
check disabled.