Converted one of the simplest ACLs: "time" (a.k.a. ACLTimeStrategy)
I did not add the "acltype name" data member to the new Acl::Time
because that would also require creating an explicit constructor,
spending memory to remember the name value, and still overriding the
typeString() method. Since the code only uses one type name with this
class -- "time" -- it feels wasteful to add all that!

We could store the acltype name as an ACL data member and set it when
Make() in Acl.cc creates a new ACL object, but since that would result
in adding a pointer to every ACL object, I did not do it either.

The current solution does not add any runtime overhead but does require
defining a typeString() method for Acl::Time and similar ACL types.