CI: Remove unnecessary test-functionality test wrappers
These workarounds are not needed for the current and future code in this
branch. Other branches have their own test-functionality.sh files that
can be used to maintain a branch-specific collection of test wrappers.

We left (now unused) has_commit_by_message() function in the script
because that function is likely to be used by future workarounds. Unlike
specific test workarounds that only apply to a subset of old code, this
and similar functions can be viewed as a reusable code "library".