Converted remaining RunnerRegistrationEntry() calls inside namespaces
We have to define the registrator function for Foo::Class inside the
global namespace: We cannot define the registrator function for
Foo::Class inside the corresponding namespace Foo because the caller of
that function would not be able to declare that registrator function --
functions cannot open namespaces (to add declarations to them). And
without such in-caller declaration, a call to that registrator function
would generate compiler "not declared in this scope" errors.