Fix memory leak when reconfiguring all-of ACL with two rules
Normally, Acl::InnerNode::add() automatically registers stored ACL nodes
for future cleanup, but when we find the second all-of rule/line with
the same name, we do not add() the newly created OrNode and have to
explicitly register it to avoid memory leaks on reconfiguration.

Leaking since all-of ACL was added in 2013 commit 6f58d7d.