Bug 5290: pure virtual call in Ftp::Client constructor
Converting `this` to CbcPointer in a constructor of an abstract class
like Ftp::Client does not work because our virtual toCbdata() method
remains pure until the child class constructor runs.

Conceptually, the bug was probably introduced in 2013 commit 434a79b,
when FTP class hierarchy grew, making Ftp::Client an abstract class, but
the trigger was recent commit 337b9aa that removed CBDATA_CLASS() from
Ftp::Client class declaration. We discovered, described, and addressed
several such bugs in that commit, but we missed this case.

XXX: This particular workaround is not based on the scheme adopted in
commit 337b9aa (because I forgotten about it and rediscovered it only
when writing this commit message). We should use that scheme instead!