fixup: Upgraded Acl::Address::AddressSource dumping to use std::ostream
Modern configuration code uses STL streams to dump configuration values.
Streams offer many (obvious?) advantages compared to printf(). The
branch had to modified nearly all of the moved lines anyway.

My primary concern here is that we are moving the buggy "autoselect"
line, creating a temptation to fix that out of scope bug in this branch.
Squid dump_foo() functions are supposed to use squid.conf syntax and
"autoselect" is not a part of that. Fixing that specific problem is
trivial, but it opens a small Pandora box: We should then print
other/all special values recognized by parseAddressToken() (i.e.
"any_addr", "no_addr", and "full_mask")  and, ideally, document them as
well. However, those special values are probably somewhat
context-sensitive -- some should not be allowed in some contexts, and
some may use context-sensitive spelling. And we will need to figure out
whether it is a good idea to actually document all of them (perhaps some
should be deprecated and/or accepted with a different spelling). Etc.

The rest of the directive dumping code and its parsing code should be
moved as well, but that code cannot be moved without significant changes
to the functions it calls. Some of those changes are actually being
implemented in parallel projects, so we are making (slow) progress.