fixup: Do not split the old tcp_outgoing_address example by new text
It is difficult to find the right position for the new text because the
existing/official documentation inserts generic rules in the middle of
its own example. I moved and detailed those rules as well, so that the
example is now located in one place, but more out-of-scope polishing is
needed here.