fixup: Reduced poor code duplication, addressing previous commit XXX
... and another old related XXX.

Also removed a stale "where to 'finish' the parsing pass?" XXX. It was
added in 2006 commit a5baffb, apparently to indicate the lack of parser
cleanup code corresponding to the then-added HttpParserInit() call.
Since our request parsing code is (or at least pretends to be)
incremental, the lack of that cleanup is not really an XXX anymore, and
the corresponding parser reset logic is documented in
Http::One::Server::parseOneRequest().

I renamed clientParseRequests() to reduce chances that some old callers
will mishandle closed connections when porting this branch changes.