Store Security::PeerOptions and Security::Context together
* Old code needs to store them together to fix Squid Bug 5293.

* New code needs to store them together because there will be many pairs
  (one per tls_outgoing_options_for_retries rule). We cannot just pair a
  few known-in-advance globals together like the (buggy) old code did.
  If we did not have to deal with GnuTLS code, we could just pass around
  Security::Context, but several low-level interfaces are already
  written to accommodate GnuTLS needs, and it would be a step back to
  rewrite them to ignore PeerOptions.

At first, I tried to store Security::PeerContext _instead_ of
Security::Options or Security::Context, but that resulted in numerous
changes unrelated to this branch feature work. To avoid that noise
explosion, I added a temporary FuturePeerContextPointer shim that
presents old (unchanged and often located in unrelated places) storage
as a single PeerContext-like view to a few new or modified places where
that conversion is needed to implement the new feature.

TODO: Update old storage (possibly in a separate Bug 5293-fixing PR to
confine all that noise!) to use PeerContext and remove the shim.