fixup: Removed unnecessary const_cast in destroy()
Conceptually, the cast is kind of expected because the visitor modifies
(and then destroys) the node. However, it is not actually necessary and
is, arguably, misleading because the Splay class itself does not promise
to keep the tree constant, even during what might look like a constant
operation to the caller: Splay::head is mutable.

There is probably a way to implement Splay tree that are meaningfully
const-correct, but that is out of scope here. Sine Splay trees should be
dropped from Squid, we may as well enjoy the positive side effects of
the current (ugly) implementation.