Avoid changing Splay::destroy() semantics w.r.t. nil cleanup function
This API change would be out of scope for this branch.

The change is also unfinished: The existing code uses empty cleanup
functions to avoid Value cleanup (e.g., DelayUserFree). If we decide to
change this API, then we should replace those empty cleanup functions
with nil pointers and adjust destroy() implementation accordingly.
However, the best way to improve destroy() API is to require proper
Value destructors, dropping the concept of a Value cleanup function
completely!