fixup: Simplified test-suite/splay.cc adjustments
I see no reason to insert that "extra" 101 node -- all inserted Values
are treated the same AFAICT. One could argue that the removal of this
101 node is an out of scope change, but I think we can argue that we are
allowed to simplify code that we have to change for other reasons.

Needless to say, the whole insert() API needs a lot of work. Even the
implementation is arguably buggy -- it assumes there can be duplicates
but does not properly cleanup them. All those problems are out of scope.