fixup: Removed misleading branch-added TODO
ErrorState is used for two very different purposes:

A) Management of error information that has not been "committed" yet
   (because some failed transactions may be retried). Error type,
   detail, and xerrno fields store error information.

B) "Rendering" of the [final] error information. BuildHttpReply() uses
   ALE, request, response_ and other fields to produce an error reply.

(A) code cannot "reflect already recorded information" as suggested by
XXX because it is the code that does recording -- ALE and friends do not
and should not have this possibly temporary/non-final info.

(B) code is the "reflection" code mentioned in XXX. Today, (B) updates
transaction records because it [correctly] assumes that the error has
been "committed" (i.e. will not be retried). BuildHttpReply() is a bad
location for those updates because some committed errors are never sent
(but are logged). Most likely, (B) should be split into two (update +
rendering) and moved away from ErrorState, but that short XXX did not
reflect these nuances/plans well enough.