fixup: Avoid optional adjustment of a DBG_CRITICAL message
The message itself and the code it is warning about are problematic.
Both will be improved/removed shortly -- the corresponding PR is waiting
for these changes to land. The term "helper client" may be confusing to
admins. It is definitely not consistent with other amdin-visible
messages/terminology. There is no need to change this message right now.
It is better to preserve this admin-visible message.