Fixed ExternalACLLookup by refactoring ACLChecklist::AsyncState
ExternalACLLookup::checkForAsync() -- the code that starts communication
with the external ACL helper -- assumed two things:

* AclMatchedName global matches this->name during ACLExternal::match().
* The ACL object name "foo" never changes during Checklist evaluation.

The first assumption might be true[1], but the second assumption is
false because reconfiguration changes the result of FindByName("foo"),
either to a new ACL object (which may have completely different
semantics!) or nil. This change can happen during full and partial
reconfiguration.

The source code comment suggested that we "have a pointer to this around
somewhere". That pointer is the "this" pointer of an ACL calling
goAsync(). It is available in the call stack two frames higher, but is
not really stored by the Checklist in any usable form AFAICT.

We could just add a second goAsync() parameter to pass that pointer to
Checklist. Doing that would require changing all the goAsync() callers
and the corresponding ACLChecklist::AsyncState API. Since that very
complex and confusing API was actually used as a basic function pointer,
we decided it is best to replace the has-to-be-changed-anyway API with a
function pointer. That plan worked well, removing quite a bit of

[1] It is difficult to be sure: AclMatchedName global was named and
added for a very different purpose, and no API enforces that invariant.