fixup: Renamed Parser::Impl to Detail_
An "internal implementation detail" namespace is often called one of
those three words[^1], abbreviated or not, with an underscore suffix or
not. Let's standardize on "Detail_" because it avoids abbreviations,
does not risk conflicting with "impl" commonly used in Pimpl idiom, and
cannot conflict with any regular Squid namespace (because of the
trailing underscore which, to many, also implies some kind of "private
use").

[^1]: https://stackoverflow.com/a/26546780