Remove AclMatchedName from ACL::ParseAclLine()
The “ugly” use of AclMatchedName global supplied ACL name to error
reporting during ACL (re)configuration. To solve this problem, we could
pass ACL::name to the reporting code. This, however, is difficult
because reporting code is located in stateless functions, such as
aclDomainCompare() or acl_ip_data::NetworkCompare() without access to
ACL::name.

Another approach (implemented in this PR) is to reuse CodeContext
concept for ACLs: the name of the ACL being parsed is appended to those
debugs() automatically by CodeContext infrastructure.