fixup: Fix and encapsulate tricky number-to-ID conversion code
Branch commit 1f283fa broke httpHdrCcStatDumper(): The new code assumed
that the given value is always positive while the caller calls that
function with a -1.0 value. That value cannot be (safely) casted to
HttpHdrCcType because the latter is based on an _unsigned_ type.

Official code did not have this bug because it casted to int instead of
HttpHdrCcType. The new branch code also does not convert out-of-range
integers into HttpHdrCcType values.

This change also places the burden of handling invalid entries (and the
choice of an "invalid" spelling, where applicable) on the caller.
Official code did not use the same spelling and, arguably, different
callers may prefer different spelling (or not to deal with invalid
entries at all because they cannot see them).