fixup: Undo introduction of references to metaHeaders() iteration
It is probably safe to use a reference (to a Note::Pointer) in the
current code, but I am a bit worried that this history-updating loop
might start modifying Adaptation::Config::metaHeaders() global,
invalidating the loop variable `h`. This is a bit paranoid, but this
branch is not about optimizing this loop iteration, so let's not take
these risks here.