Fix proxy processing time for pinned connections
[http]::<tt was empty for bumped at 2/3 step configurations
(both fwd and tunnel) because HierarchyLogEntry::startPeerClock()
was not called for pinned connections.

Also adjusted [http]::<tt definition to account for the peer lookup
stage (which runs in parallel with connect attempts).

Also fixed [http]::<tt for peek-and-splice step 2/3 configurations when
a TLS error initiates switchToTunnel(). In this case, the logged
time covered only the interval up to the failed TLS attempt and
did not include the rest of the tunnel communication.