Adjusted updateReply() and updateAle(): The last update() wins
This approach allows the caller to overwrite "default" (e.g.,
ALE-derived or parent-set values) with more specific ones.

In case of updateAle(), this approach also avoids surprises when the
first updateAle() call (not shown below) did not have ALE::reply set and
the second updateAle() call had ALE::reply set but could not apply it to
the checklist.

    ... checklist updated in parent class or some "fill" method ...
    ale->reply = myReply; // we are the only code that knows myReply
    checklist.updateAle(ale); // the second (ignored) updateAle() call
    // checklist.reply() crashes here because there is no reply set!

Also call updateReply(r) _after_ updateAle(ale) in case r is different
from ale->reply (and both are not nil).