Renamed and polished aclGetDenyInfoPage() further
All callers were supplying the same first argument that is unlikely to
change in the foreseeable future.

Renaming emphasizes that the function may not find what the caller is
looking for (if we did not have ERR_NONE, the function would return
std::optional). And if we rename, we should supply proper description.

I did not fix redirect_allowed name and debugs(). I am not sure what the
correct behavior is. Fixing this behavior deserves a dedicated PR IMO.
Other aclGetDenyInfoPage() changes in this PR do not affect behavior.