Clarify ALE::tcpClient scope; undo acceptedClientConnection changes
AccessLogEntry::acceptedClientConnection was added based on
https://github.com/measurement-factory/squid/pull/245#discussion_r1478608541

I believe that framework reasoning and long-term plans were overall
correct except for tcpClient interpretation/plans. And with that
interpretation corrected, there is a better way to address the immediate
problem at hand than adding acceptedClientConnection hack to ALE.

I described (future) ALE::tcpClient as eventually representing "the
primary connection of the corresponding transaction. ... For HTTP
Squid-peer transactions, that is the connection opened by Squid". Given
tcpClient use in current Squid code (where ICAP does not set it at all
and all users [can be interpreted as if they] extract
accepted-connection information like %ui from it), that was not the best
model. The field should be viewed as representing the connection
accepted by Squid. It is a MasterXaction-level info (duplicated in ALE).
A connection opened by FwdState or by an ICAP RESPMOD transaction should
not be stored in tcpClient, even if/when those transactions gain their
own ALEs (those future ALE-like types might have something like a
primaryConnection field for their connections).

If ALE::tcpClient represents the connection accepted by Squid (at one of
its proto_ports), then ICAP code should use that field to extract Ident
user identity (for the purposes of %ui logging and ident_regex matching)
because all related %codes and ACLs are defined in terms of the user
opening a connection to a Squid port (and _not_ the user of a connection
Squid opens to an HTTP peer or an ICAP service!).