fixup: ~Restore code restoring ALE IDENT info
I do not know whether this code is needed in some cases, but it feels
wrong to silently remove it, especially when a comment at the beginning
of ACLFilledChecklist::verifyAle() implies that this code may be useful.

Code details have changed because ALE no longer stores cache.rfc931
directly, but tcpClient in new code is needed to extract that same Ident
info.