fixup: Removed misleading and out-of-scope/place XXX
It is fine for configuration parsing code to call FindByName(): We store
new ACLs in Config that FindByName() is checking.

Code like aclIsProxyAuth() caller (e.g, dyny_info) does suffer from this
XXX: That code may extract information from an irrelevant ACL that
happens to share the name with an old relevant ACL that was reconfigured
away. However, that problem was not introduced by this branch changes,
and is arguably not FindByName() fault -- ACLChecklist itself should
probably relay either the last checked ACL Node (C++ object, not just
its name) or its "is proxy auth" info.